requirements.txt
numpy
psutil
PyAudio
pyserial
sounddevice
structlog
requests
asyncio
chardet
colorama
ordered-set
nuitka<=2.6.9
pyinstaller
websocket-client
fastapi[standard]
uvicorn[standard]

# Development and test dependencies
autopep8
black
isort
pycodestyle
pytest
pytest-cov
pytest-cover
pytest-coverage
pytest-rerunfailures
pick
sqlalchemy